Create AppShell
