+++ frontend/app-development/layout/App.tsx
@@ -2,26 +2,8 @@ import React, { useEffect } from 'react';
 import postMessages from 'app-shared/utils/postMessages';
 import { Outlet, matchPath, useLocation } from 'react-router-dom';
 import classes from './App.module.css';
-import i18next from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import nb from '../../language/src/nb.json';
-import en from '../../language/src/en.json';
-import { DEFAULT_LANGUAGE } from 'app-shared/constants';
 import { useRepoStatusQuery } from 'app-shared/hooks/queries';
 import { appContentWrapperId } from '@studio/testing/testids';
-
-i18next.use(initReactI18next).init({
-  lng: DEFAULT_LANGUAGE,
-  resources: {
-    nb: { translation: nb },
-    en: { translation: en },
-  },
-  fallbackLng: 'nb',
-  react: {
-    transSupportBasicHtmlNodes: true,
-    transKeepBasicHtmlNodesFor: ['em'],
-  },
-});
 
 export function App() {
   const { pathname } = useLocation();
