+++ frontend/app-preview/index.tsx
@@ -4,19 +4,17 @@ import { PreviewApp } from './src/PreviewApp';
 import { BrowserRouter } from 'react-router-dom';
 import { PREVIEW_BASENAME } from 'app-shared/constants';
 import { PreviewConnectionContextProvider } from 'app-shared/providers/PreviewConnectionContext';
-import { ServicesContextProvider } from 'app-shared/contexts/ServicesContext';
-import * as queries from 'app-shared/api/queries';
-import * as mutations from 'app-shared/api/mutations';
+import { AppShell } from 'app-shared/AppShell';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
 
 root.render(
-  <ServicesContextProvider {...queries} {...mutations}>
+  <AppShell>
     <PreviewConnectionContextProvider>
       <BrowserRouter basename={PREVIEW_BASENAME}>
         <PreviewApp />
       </BrowserRouter>
     </PreviewConnectionContextProvider>
-  </ServicesContextProvider>,
+  </AppShell>,
 );
